eventFormatter works w event id
