+++ package.json
@@ -12,6 +12,7 @@
     "@fortawesome/fontawesome-svg-core": "^6.7.2",
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
+    "@google/generative-ai": "^0.24.0",
     "@react-google-maps/api": "^2.20.6",
     "@types/leaflet": "^1.9.16",
     "firebase": "^11.4.0",
